Add inMemoryTestApplication factory method for testing

+++ src/main/java/pl/zycienakodach/pragmaticflights/modules/pricing/api/commands/DefineRegularPrice.java
@@ -1,7 +1,9 @@
 package pl.zycienakodach.pragmaticflights.modules.pricing.api.commands;
 
+import java.math.BigDecimal;
+
 public record DefineRegularPrice(
     String flightCourseId,
-    double priceInEuro
+    BigDecimal priceInEuro
 ) {
 }
